examples/xgboost-ezkl/requirements.txt
ezkl
xgboost
scikit-learn
numpy
onnx
onnxruntime
hummingbird-ml
torch
